+++ views/searchresults.php
@@ -1,0 +1,2 @@
+<?php $title = "Zoekresultaten | World Wide Importers";
+include("../templates/header.php");?>